feat: devision components, preloader, refactor code, based on #53
