+++ src/assets/styles/animation.scss
@@ -88,3 +88,12 @@
     transform: translate(1px, -2px) rotate(-1deg);
   }
 }
+
+@keyframes spinner {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
