+++ src/components/reducers/cart.ts
@@ -1,5 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { CartItem, CartItemArg, Product } from '@/components/types/types';
+import { v4 as uuidv4 } from 'uuid';
 
 const initialState: CartItem[] = [];
 
@@ -9,11 +10,14 @@ const cartSlice = createSlice({
   reducers: {
     addProductToCart: (state, { payload: cartItem }: PayloadAction<CartItemArg>): CartItem[] => {
       const item = state.find((item) => item.product.id === cartItem.product.id);
+      const cartID = uuidv4();
+
       if (!item) {
         return [
           ...state,
           {
-            id: state.length + 1,
+            cartID,
+            itemNumber: state.length + 1,
             product: cartItem.product,
             quantity: cartItem.quantity,
           },
@@ -33,11 +37,14 @@ const cartSlice = createSlice({
       { payload: cartItem }: PayloadAction<CartItemArg>
     ): CartItem[] => {
       const item = state.find((item) => item.product.id === cartItem.product.id);
+      const cartID = uuidv4();
+
       if (!item) {
         return [
           ...state,
           {
-            id: state.length + 1,
+            cartID,
+            itemNumber: state.length + 1,
             product: cartItem.product,
             quantity: cartItem.quantity,
           },
@@ -54,7 +61,7 @@ const cartSlice = createSlice({
         if (item.quantity <= 1) {
           return state
             .filter((stateItem) => stateItem !== item)
-            .map((p, i) => ({ ...p, id: i + 1 }));
+            .map((p, i) => ({ ...p, itemNumber: i + 1 }));
         }
         item.quantity -= 1;
       }
@@ -65,8 +72,15 @@ const cartSlice = createSlice({
     removeProductFromCartAll: (state, { payload: product }: PayloadAction<Product>): CartItem[] => {
       const item = state.find((item) => item.product.id === product.id);
       if (item) {
-        return state.filter((stateItem) => stateItem !== item).map((p, i) => ({ ...p, id: i + 1 }));
+        return state
+          .filter((stateItem) => stateItem !== item)
+          .map((p, i) => ({ ...p, itemNumber: i + 1 }));
       }
+      return state;
+    },
+
+    removeAllProductFromCart: (state): CartItem[] => {
+      state = [];
       return state;
     },
   },
@@ -77,5 +91,6 @@ export const {
   setProductQuantityInCart,
   removeProductFromCart,
   removeProductFromCartAll,
+  removeAllProductFromCart,
 } = cartSlice.actions;
 export default cartSlice.reducer;
